+++ app/models/slice/event_design.rb
@@ -2,19 +2,21 @@
 
 # Tracks a series of events on a design.
 class Slice::EventDesign
-  attr_accessor :subject_event, :json, :sheets, :design_id, :design_name, :event_id
+  attr_accessor :subject_event, :json, :sheets, :design_id, :design_name, :event_id, :design_slug, :event_slug
 
   def initialize(json, subject_event)
     @subject_event = subject_event
     @json = json
-    @design_id = json.dig("design", "id")
+    @event_id = json.dig("event", "actual_id")
+    @event_slug = json.dig("event", "slug").presence || @event_id
+    @design_id = json.dig("design", "actual_id")
+    @design_slug = json.dig("design", "slug").presence || @design_id
     @design_name = json.dig("design", "name")
-    @event_id = json.dig("event", "id")
     @sheets = load_sheets(json["sheets"])
   end
 
   def valid?
-    @event_id.present? && @design_id.present? && @design_name.present?
+    @event_slug.present? && @design_slug.present? && @design_name.present?
   end
 
   def complete?(subject)